backend/pyproject.toml
[project]
name = "deep-memory-explorer"
version = "0.1.0"
description = "Knowledge graph IDE for HiveMindDB — search, visualize, chat, benchmark"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "mcp>=1.0.0",
    "websockets>=13.0",
    "sse-starlette>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
